requirements.txt
asgiref==3.3.4
certifi==2021.5.30
Django==3.2.4
django-haystack==3.0
elasticsearch==5.0.0
gunicorn==20.1.0
mysqlclient==2.0.3
pytz==2021.1
sqlparse==0.4.1
urllib3==1.26.5
